move to /streams
