+++ tools/cabana/streams/livestream.cc
@@ -94,3 +94,40 @@ void LiveStream::pause(bool pause) {
   pause_ = pause;
   emit paused();
 }
+
+// OpenDeviceWidget
+
+#include <QButtonGroup>
+#include <QFormLayout>
+#include <QRadioButton>
+#include <QRegularExpression>
+#include <QRegularExpressionValidator>
+
+OpenDeviceWidget::OpenDeviceWidget(AbstractStream **stream, QWidget *parent) : AbstractOpenStreamWidget(stream, parent) {
+  QRadioButton *msgq = new QRadioButton(tr("MSGQ"));
+  QRadioButton *zmq = new QRadioButton(tr("ZMQ"));
+  ip_address = new QLineEdit(this);
+  ip_address->setPlaceholderText(tr("Enter the IpAddress"));
+  QString ip_range = "(?:[0-1]?[0-9]?[0-9]|2[0-4][0-9]|25[0-5])";
+  QString pattern("^" + ip_range + "\\." + ip_range + "\\." + ip_range + "\\." + ip_range + "$");
+  QRegularExpression re(pattern);
+  ip_address->setValidator(new QRegularExpressionValidator(re, this));
+
+  group = new QButtonGroup(this);
+  group->addButton(msgq, 0);
+  group->addButton(zmq, 1);
+
+  QFormLayout *form_layout = new QFormLayout(this);
+  form_layout->addRow(msgq);
+  form_layout->addRow(zmq, ip_address);
+  QObject::connect(group, qOverload<QAbstractButton *, bool>(&QButtonGroup::buttonToggled), [=](QAbstractButton *button, bool checked) {
+    ip_address->setEnabled(button == zmq && checked);
+  });
+  zmq->setChecked(true);
+}
+
+bool OpenDeviceWidget::open() {
+  QString ip = ip_address->text().isEmpty() ? "127.0.0.1" : ip_address->text();
+  *stream = new LiveStream(qApp, group->checkedId() == 0 ? "" : ip);
+  return true;
+}
