+++ tools/cabana/streams/replaystream.cc
@@ -50,3 +50,97 @@ void ReplayStream::pause(bool pause) {
   replay->pause(pause);
   emit(pause ? paused() : resume());
 }
+
+
+// OpenReplayWidget
+
+#include <QDialogButtonBox>
+#include <QLabel>
+#include <QFileDialog>
+#include <QGridLayout>
+#include <QMessageBox>
+#include <QPushButton>
+
+#include "common/prefix.h"
+
+static std::unique_ptr<OpenpilotPrefix> op_prefix;
+
+OpenReplayWidget::OpenReplayWidget(AbstractStream **stream, QWidget *parent) : AbstractOpenStreamWidget(stream, parent) {
+  // TODO: get route list from api.comma.ai
+  QGridLayout *grid_layout = new QGridLayout();
+  grid_layout->addWidget(new QLabel(tr("Name")), 0, 0);
+  grid_layout->addWidget(route_edit = new QLineEdit(this), 0, 1);
+  route_edit->setPlaceholderText(tr("Enter remote route name or click browse to select a local route"));
+  auto file_btn = new QPushButton(tr("Browse..."), this);
+  grid_layout->addWidget(file_btn, 0, 2);
+
+  grid_layout->addWidget(new QLabel(tr("Video")), 1, 0);
+  grid_layout->addWidget(choose_video_cb = new QComboBox(this), 1, 1);
+  QString items[] = {tr("No Video"), tr("Road Camera"), tr("Wide Road Camera"), tr("Driver Camera"), tr("QCamera")};
+  for (int i = 0; i < std::size(items); ++i) {
+    choose_video_cb->addItem(items[i]);
+  }
+  choose_video_cb->setCurrentIndex(1);  // default is road camera;
+
+  QVBoxLayout *main_layout = new QVBoxLayout(this);
+  main_layout->addLayout(grid_layout);
+  setMinimumWidth(550);
+
+  QObject::connect(file_btn, &QPushButton::clicked, [=]() {
+    QString dir = QFileDialog::getExistingDirectory(this, tr("Open Local Route"), settings.last_route_dir);
+    if (!dir.isEmpty()) {
+      route_edit->setText(dir);
+      settings.last_route_dir = QFileInfo(dir).absolutePath();
+    }
+  });
+}
+
+bool OpenReplayWidget::open() {
+  QString route = route_edit->text();
+  QString data_dir;
+  if (int idx = route.lastIndexOf('/'); idx != -1) {
+    data_dir = route.mid(0, idx + 1);
+    route = route.mid(idx + 1);
+  }
+
+  failed_to_load = true;
+  bool is_valid_format = Route::parseRoute(route).str.size() > 0;
+  if (!is_valid_format) {
+    QMessageBox::warning(nullptr, tr("Warning"), tr("Invalid route format: '%1'").arg(route));
+  } else {
+    // TODO: Remove when OpenpilotPrefix supports ZMQ
+#ifndef __APPLE__
+    op_prefix.reset(new OpenpilotPrefix());
+#endif
+    uint32_t flags[] = {REPLAY_FLAG_NO_VIPC, REPLAY_FLAG_NONE, REPLAY_FLAG_ECAM, REPLAY_FLAG_DCAM, REPLAY_FLAG_QCAMERA};
+    ReplayStream *replay_stream = *stream ? (ReplayStream *)*stream : new ReplayStream(qApp);
+    failed_to_load = !replay_stream->loadRoute(route, data_dir, flags[choose_video_cb->currentIndex()]);
+    if (failed_to_load) {
+      if (replay_stream != *stream) {
+        delete replay_stream;
+      }
+      QMessageBox::warning(nullptr, tr("Warning"), tr("Failed to load route: '%1'").arg(route));
+    } else {
+      *stream = replay_stream;
+    }
+  }
+  return !failed_to_load;
+}
+
+
+// OpenRouteDiloag
+
+OpenRouteDialog::OpenRouteDialog(QWidget *parent) : QDialog(parent) {
+  setWindowTitle(tr("Open Route"));
+  QVBoxLayout *main_layout = new QVBoxLayout(this);
+  main_layout->addWidget(replay_widget = new OpenReplayWidget(&can, this));
+  auto btn_box = new QDialogButtonBox(QDialogButtonBox::Open | QDialogButtonBox::Cancel);
+  main_layout->addWidget(btn_box);
+
+  QObject::connect(btn_box, &QDialogButtonBox::rejected, this, &QDialog::reject);
+  QObject::connect(btn_box, &QDialogButtonBox::accepted, [=]() {
+    if (replay_widget->open()) {
+      accept();
+    }
+  });
+}
