+++ tools/cabana/streams/replaystream.h
@@ -31,3 +31,30 @@ private:
   std::unique_ptr<Replay> replay = nullptr;
   std::set<int> processed_segments;
 };
+
+class OpenReplayWidget : public AbstractOpenStreamWidget {
+  Q_OBJECT
+
+public:
+  OpenReplayWidget(AbstractStream **stream, QWidget *parent);
+  bool open() override;
+  QString title() override { return tr("Replay"); }
+  inline bool failedToLoad() const { return failed_to_load; }
+
+private:
+  QLineEdit *route_edit;
+  QComboBox *choose_video_cb;
+  bool failed_to_load = false;
+};
+
+
+class OpenRouteDialog : public QDialog {
+  Q_OBJECT
+
+ public:
+  OpenRouteDialog(QWidget *parent);
+  inline bool failedToLoad() const { return replay_widget->failedToLoad(); }
+
+ private:
+  OpenReplayWidget *replay_widget;
+};
